+++ lib/neo4j/shared/persistence.rb
@@ -84,7 +84,7 @@ module Neo4j::Shared
 
     def touch
       fail 'Cannot touch on a new record object' unless persisted?
-      update_attribute(:updated_at, Time.now) if attributes.key?('updated_at')
+      update_attribute!(:updated_at, Time.now) if respond_to?(:updated_at=)
     end
 
     # Returns +true+ if the record is persisted, i.e. it's not a new record and it was not destroyed